- Remove warnings - Remove `USE_REGISTER`. Now it is used by default.

+++ app/app_cmdbox.c
@@ -212,7 +212,7 @@ static size_t app_cmdbox_parse_search(const char *cmd, const tAppCmdboxDatabaseL
   const tAppCmdboxDatabaseListUnit *p_matched_cmd = NULL;
   size_t                           cursor         = 0;
 
-  if (isalpha(cmd[0]) && cmd[0] >= 'A') {
+  if (isalpha((int)cmd[0]) && cmd[0] >= 'A') {
     const tAppCmdboxDatabaseList *cmd_database_list = CMD_TABLE[cmd[cursor]-'A'];
     /* Search if command was supported */
     const char *ptr   = NULL;
